[vmfs-bal] separate common functions into a helper module
this new helper module will be merged into the existing ones
later after merging this branch
